Lab4: Green

+++ src/exercise5.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 
 int main() {
-    int A[10][3];
-    int B[3][10];
-    int result[10][10];
+    float A[10][3];
+    float B[3][10];
+    float result[10][10];
     
     for (int i = 0; i < 10; i++) {
         for (int j = 0; j < 3; j++) {
-            scanf("%d", &A[i][j]);
+            scanf("%f", &A[i][j]);
         }
     }
     
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 10; j++) {
-            scanf("%d", &B[i][j]);
+            scanf("%f", &B[i][j]);
         }
     }
     
@@ -28,7 +28,7 @@ int main() {
     
     for (int i = 0; i < 10; i++) {
         for (int j = 0; j < 10; j++) {
-            printf("%d ", result[i][j]);
+            printf("%.2f ", result[i][j]);
         }
     }
     printf("\n");
